requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
click==8.1.8
contourpy==1.3.2
cron-descriptor==1.4.5
cryptography==44.0.2
cycler==0.12.1
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-timezone-field==7.1
Django==5.1.7
djangorestframework==3.15.2
flask-cors==5.0.1
Flask==3.1.0
fonttools==4.58.0
idna==3.10
joblib==1.5.0
kiwisolver==1.4.8
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.50
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
seaborn==0.13.2
sendgrid==6.11.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
threadpoolctl==3.6.0
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
xgboost==3.0.0